+++ config/generators.php
@@ -6,11 +6,22 @@ return [
      * have to implement Webfactor\Laravel\Generators\Contracts\MakeServiceInterface.
      */
     'services' => [
-        Webfactor\Laravel\Generators\Services\MakeBackpackCrudModelService::class,
-        Webfactor\Laravel\Generators\Services\MakeBackpackCrudControllerService::class,
-        Webfactor\Laravel\Generators\Services\MakeBackpackCrudRequestService::class,
-        Webfactor\Laravel\Generators\Services\MakeMigrationService::class,
-        Webfactor\Laravel\Generators\Services\MakeFactoryService::class,
-        Webfactor\Laravel\Generators\Services\MakeSeederService::class,
+        Webfactor\Laravel\Generators\Services\MigrationService::class,
+        Webfactor\Laravel\Generators\Services\LanguageService::class,
+        Webfactor\Laravel\Generators\Services\BackpackCrudModelService::class,
+        Webfactor\Laravel\Generators\Services\BackpackCrudControllerService::class,
+        Webfactor\Laravel\Generators\Services\BackpackCrudRequestService::class,
+        Webfactor\Laravel\Generators\Services\FactoryService::class,
+        Webfactor\Laravel\Generators\Services\SeederService::class,
+        Webfactor\Laravel\Generators\Services\RouteService::class,
+        Webfactor\Laravel\Generators\Services\OpenIdeService::class,
+    ],
+
+    /*
+     * Recipe classes for opening all generated files directly in IDE if the option --ide={key}
+     * is used. Have to implement Webfactor\Laravel\Generators\Contracts\OpenIdeInterface.
+     */
+    'ides'     => [
+        'pstorm' => Webfactor\Laravel\Generators\Recipes\PhpStormOpener::class,
     ],
 ];