+++ CHANGELOG.md
@@ -21,6 +21,19 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ### Security
 - Nothing
 
+## 2.0.0 - 2018-03-23
+
+### Added
+- `LanguageService` generates models.php translation file (if not exists) and fill singular/plural translation
+- `OpenIdeService` opens all generated file with PhpStorm if command is called with `--ide={ide}`
+- `RouteService` adds Backpack Crud route to admin.php
+
+### Changed
+
+- `BackpackCrudModelService` will fill `$fillable` in Model automatically from scheme (if given)
+- `BackpackCrudRequestService` will fill `rules()` in Request automatically from scheme (if given)
+- `BackpackCrudControllerService` will add CrudFields and CrudColumns in Controller automatically from scheme (if given) - very rudimentary for now, more functionality planned
+
 ## 1.2.0 - 2018-03-18
 
 ### Added